third_party/WebKit/LayoutTests/inspector/console/console-uncaught-exception.html

<html>
<head>
<script src="../../http/tests/inspector/inspector-test.js"></script>
<script src="../../http/tests/inspector/console-test.js"></script>
<script>

function test() {
  ConsoleTestRunner.addConsoleViewSniffer(addMessage, true);
  TestRunner.evaluateInPage('loadIframe()');
  function addMessage(viewMessage) {
    if (viewMessage.element().deepTextContent().indexOf('setTimeout') !== -1)
      ConsoleTestRunner.expandConsoleMessages(onExpanded);
  }

  function onExpanded() {
    ConsoleTestRunner.dumpConsoleMessages();
    TestRunner.completeTest();
  }
}

function loadIframe()
{
    var iframe = document.createElement("iframe");
    iframe.src = "resources/uncaught-in-iframe.html";
    document.body.appendChild(iframe);
}

</script>
</head>

<body onload="runTest()">
<p>
Tests that uncaught exceptions are logged into console.<a href="https://bugs.webkit.org/show_bug.cgi?id=47250">Bug 47250.</a>
</p>

</body>
</html>
